Add Playlist table to Database, Add PlaylistFragment

+++ wandm/app/src/main/java/com/wandm/models/playlist/Playlist.kt
@@ -2,7 +2,7 @@ package com.wandm.models.playlist
 
 class Playlist {
 
-    val id: Long
+    val id: Int
     val name: String
     val songCount: Int
 
@@ -12,7 +12,7 @@ class Playlist {
         this.songCount = -1
     }
 
-    constructor(_id: Long, _name: String, _songCount: Int) {
+    constructor(_id: Int, _name: String, _songCount: Int) {
         this.id = _id
         this.name = _name
         this.songCount = _songCount
